Extracts time functions into own module
